+++ e2e/browser/test-app/pages/_app.tsx
@@ -19,19 +19,11 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-export { default as issueVerifiableCredential } from "./issue/issue";
-export type { Iri, JsonLd, VerifiableCredential } from "./common/common";
-export {
-  isVerifiableCredential,
-  isVerifiablePresentation,
-  getVerifiableCredential,
-  getVerifiableCredentialApiConfiguration,
-} from "./common/common";
-export { default as getVerifiableCredentialAllFromShape } from "./lookup/derive";
-export { query } from "./lookup/query";
-export type {
-  QueryByExample,
-  VerifiablePresentationRequest,
-} from "./lookup/query";
-export { revokeVerifiableCredential } from "./revoke/revoke";
-export { isValidVc, isValidVerifiablePresentation } from "./verify/verify";
+import React from "react";
+
+/* eslint react/prop-types: 0, react/jsx-props-no-spreading: 0 */
+function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;